+++ mapping_suite_sdk/mapping_package_v3/adapters/mp_v3L_package_saver.py
@@ -7,7 +7,7 @@ import logging
 
 from pymongo import MongoClient
 
-from mapping_suite_sdk.core.adapters.gridfs_repository import GridFSPackageRepository
+from mapping_suite_sdk.core.adapters.package_repository import PackageRepository
 from mapping_suite_sdk.core.adapters.tracer import traced_class
 from mapping_suite_sdk.mapping_package_v3.models.mapping_package_v3_lightweight import (
     MappingPackageV3Lightweight
@@ -41,7 +41,7 @@ class MappingPackageV3LightweightSaver:
         if not mongo_client:
             raise ValueError("MongoDB client must be provided")
 
-        repository = GridFSPackageRepository[MappingPackageV3Lightweight](
+        repository = PackageRepository[MappingPackageV3Lightweight](
             model_class=MappingPackageV3Lightweight,
             mongo_client=mongo_client,
             database_name=database_name,
